requirements.txt
yacs
torchvision
tqdm
